+++ src/store/actions/unlockWallet.js
@@ -1,31 +1,10 @@
 import { encrypt, decrypt, decryptLegacy } from '../../utils/crypto'
-import { Client } from '@liquality/client'
-import { EthereumRpcProvider } from '@liquality/ethereum-rpc-provider'
+import { getLegacyRskBalance } from '../utils'
 
 export const unlockWallet = async ({ commit, state, dispatch }, { key }) => {
   let wallets = await decrypt(state.encryptedWallets, key, state.keySalt)
 
-  const accKeys = Object.keys(state.accounts);
-
-  let addresses = [];
-
-  for(let i = 0; i < accKeys.length; i++) {
-    const currentAccs = currentState.accounts[accKeys[i]].mainnet;
-    
-    for(let k = 0; k < currentAccs.length; k++) {
-      if(currentAccs[k].chain === 'rsk') {
-        addresses.push(...currentAccs[k].addresses)
-      }
-    }
-  }
-
-  const client = new Client()
-    .addProvider(
-      new EthereumRpcProvider({ uri: 'https://public-node.rsk.co' })
-    );
-  
-  const balance = await client._chain.getBalance(addresses)
-  
+  const balance = await getLegacyRskBalance(state.accounts)  
   // Migration to new encryption method
   // TODO: to be removed
   if (!wallets) {
