+++ src/store/migrations.js
@@ -3,8 +3,7 @@ import buildConfig from '../build.config'
 import { accountCreator, getNextAccountColor } from '@/utils/accounts'
 import { chains, assets as cryptoassets } from '@liquality/cryptoassets'
 import { v4 as uuidv4 } from 'uuid'
-import { Client } from '@liquality/client'
-import { EthereumRpcProvider } from '@liquality/ethereum-rpc-provider'
+import { getLegacyRskBalance } from './utils'
 
 const migrations = [
   { // Merely sets up the version
@@ -216,33 +215,14 @@ const migrations = [
   { // rskLegacyDerivation
     version: 11,
     migrate: async (state) => {
-      const accKeys = Object.keys(state.accounts);
-
-      if (!accKeys?.length) {
+      if (!Object.keys(state.accounts)?.length) {
         return {
           ...state,
         }
       }
 
-      let addresses = [];
+      const balance = await getLegacyRskBalance(state.accounts) 
 
-      for(let i = 0; i < accKeys.length; i++) {
-        const currentAccs = currentState.accounts[accKeys[i]].mainnet;
-        
-        for(let k = 0; k < currentAccs.length; k++) {
-          if(currentAccs[k].chain === 'rsk') {
-            addresses.push(...currentAccs[k].addresses)
-          }
-        }
-      }
-
-      const client = new Client()
-        .addProvider(
-          new EthereumRpcProvider({ uri: 'https://public-node.rsk.co' })
-        );
-      
-      const balance = await client._chain.getBalance(addresses)
-      
       return {
         ...state,
         rskLegacyDerivation: !balance.toNumber()
