+++ src/store/utils.js
@@ -5,6 +5,8 @@ import { assets as cryptoassets } from '@liquality/cryptoassets'
 import { BitcoinNetworks } from '@liquality/bitcoin-networks'
 import { EthereumNetworks } from '@liquality/ethereum-networks'
 import { NearNetworks } from '@liquality/near-networks'
+import { Client } from '@liquality/client'
+import { EthereumRpcProvider } from '@liquality/ethereum-rpc-provider'
 
 export const CHAIN_LOCK = {}
 
@@ -44,6 +46,29 @@ export const unlockAsset = key => {
 }
 
 const COIN_GECKO_API = 'https://api.coingecko.com/api/v3'
+
+export const getLegacyRskBalance = async(accounts) => {
+  const walletIds = Object.keys(accounts);
+  
+  let addresses = [];
+
+  walletIds.forEach((wallet, idx) => {
+    const walletAccounts = accounts[wallet[idx]].mainnet;
+
+    walletAccounts.forEach(account => {
+      if(account.chain === 'rsk') {
+        addresses.push(...currentAccs[k].addresses)
+      }
+    })
+  })
+
+  const client = new Client()
+    .addProvider(
+      new EthereumRpcProvider({ uri: 'https://public-node.rsk.co' })
+    );
+  
+  return await client._chain.getBalance(addresses)
+}
 
 export async function getPrices (baseCurrencies, toCurrency) {
   const coindIds = baseCurrencies.filter(currency => cryptoassets[currency]?.coinGeckoId)
